Plots: Extract PlotCanvasBase
